update examples and remove validation tests
