+++ example/generators/generators.cpp
@@ -47,10 +47,15 @@ public:
     //    capacitance: 0.01 F/m² [default]
     //    synapses: 1 * expsyn
     arb::util::unique_any get_cell_description(cell_gid_type gid) const override {
-        arb::cable_cell c;
+        arb::sample_tree tree;
+        double r = 18.8/2.0; // convert 18.8 μm diameter to radius
+        tree.append({{0,0,0,r}, 1});
 
-        c.add_soma(18.8/2.0); // convert 18.8 μm diameter to radius
-        c.soma()->add_mechanism("pas");
+        arb::label_dict d;
+        d.set("soma", arb::reg::tagged(1));
+
+        arb::cable_cell c(tree, d);
+        c.paint("soma", "pas");
 
         // Add one synapse at the soma.
         // This synapse will be the target for all events, from both
