+++ sup/include/sup/json_params.hpp
@@ -1,3 +1,5 @@
+#pragma once
+
 #include <array>
 #include <exception>
