--- test/validation/make_image.sh
@@ -1,5 +1,0 @@
-dot cell.dot -Tpdf -o cell.pdf
-# uses display from imagemagik
-display cell.pdf
-# or just call open on os x
-#open cell.pdf
